Move non-scss related code to root directory
* Move ScssLint::File to SourceFile
* Move ScssLint::SystemCall to SystemCall
