+++ spec/lib/source_file.rb
@@ -1,13 +1,13 @@
 require "spec_helper"
-require "scss_lint/file"
+require "source_file"
 
-describe ScssLint::File do
+describe SourceFile do
   describe ".in_tmpdir" do
     it "writes each file into the tmpdir and yields the dir" do
-      file_one = ScssLint::File.new("file-one.scss", "some content")
-      file_two = ScssLint::File.new("file-two.scss", "some content")
+      file_one = SourceFile.new("file-one.scss", "some content")
+      file_two = SourceFile.new("file-two.scss", "some content")
 
-      ScssLint::File.in_tmpdir(file_one, file_two) do |dir|
+      SourceFile.in_tmpdir(file_one, file_two) do |dir|
         expect(Dir.entries(dir)).to include("file-one.scss", "file-two.scss")
       end
     end
@@ -16,7 +16,7 @@ describe ScssLint::File do
   describe "#write_to_dir" do
     it "writes files in the given directory" do
       Dir.mktmpdir do |dir|
-        file = ScssLint::File.new("file.scss", "")
+        file = SourceFile.new("file.scss", "")
 
         file.write_to_dir(dir)
