+++ spec/lib/system_call_spec.rb
@@ -1,10 +1,10 @@
 require "spec_helper"
-require "scss_lint/system_call"
+require "system_call"
 
-describe ScssLint::SystemCall do
+describe SystemCall do
   context "running valid command" do
     it "captures stdout" do
-      system = ScssLint::SystemCall.new
+      system = SystemCall.new
 
       output = system.call("printf 'Hello World!'")
 
@@ -12,7 +12,7 @@ describe ScssLint::SystemCall do
     end
 
     it "captures stderr" do
-      system = ScssLint::SystemCall.new
+      system = SystemCall.new
 
       output = system.call("printf 'Hello World!' 1>&2")
 
@@ -20,7 +20,7 @@ describe ScssLint::SystemCall do
     end
 
     it "captures and concats stdout and stderr" do
-      system = ScssLint::SystemCall.new
+      system = SystemCall.new
 
       command = "printf 'Hello Stdout!' && printf 'Hello Stderr!' 1>&2"
       output = system.call(command)
@@ -31,17 +31,17 @@ describe ScssLint::SystemCall do
 
   context "running an invalid command" do
     it "raises an error" do
-      system = ScssLint::SystemCall.new
+      system = SystemCall.new
 
       expect { system.call("printf") }.
         to raise_error(
-          ScssLint::SystemCall::NonZeroExitStatusError,
+          SystemCall::NonZeroExitStatusError,
           "Command: 'printf'",
         )
     end
 
     it "stores the command's output on the error" do
-      system = ScssLint::SystemCall.new
+      system = SystemCall.new
 
       expect { system.call("printf") }.to raise_error do |error|
         expect(error.output).
